Change to move the dummy_pycompss.py module to the utils dir so that it can be used by all pipelines
