+++ tools_demos/simpleTool1.py
@@ -27,8 +27,8 @@ except ImportError:
     print("[Warning] Cannot import \"pycompss\" API packages.")
     print("          Using mock decorators.")
 
-    from dummy_pycompss import FILE_IN, FILE_OUT
-    from dummy_pycompss import task
+    from utils.dummy_pycompss import FILE_IN, FILE_OUT
+    from utils.dummy_pycompss import task
 
 from basic_modules.metadata import Metadata
 from basic_modules.tool import Tool
@@ -77,7 +77,7 @@ class SimpleTool1(Tool):
         taskStatus = self.inputPlusOne(input_files["input"],
                                        output_files["output"])
         logger.info("SimpleTool1: task inputPlusOne done")
-        
+
         if taskStatus:
             logger.info("SimpleTool1: run successful")
             return ({"output": output_files["output"]},
